Create NearestNeighbor.java
This is the nearest neighbor algorithm, which finds a low distance solution to the problem by picking a random node as the starting node, and then continuously choosing the node closest to the previously chosen node. This is perhaps the simplest solution to the problem, but often yields suboptimal results.