Create NearestInsertion.java
The nearest insertion algorithm for TSP, it picks two points, and then picks the point that is closest to two adjacent points on the path, and it removes the connection between the two points but adds the connection for the new points to the two previous points. This is continued until all points are included on the path.